Replace mock function with rpc call
